Header in GetPRListActivity and fine refactor
