+++ app/src/main/java/com/test/procore/fetchgithubdata/activities/GetPRListActivity.java
@@ -1,14 +1,13 @@
-package com.test.procore.fetchgithubdata;
+package com.test.procore.fetchgithubdata.activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
-import android.view.ViewDebug;
+import android.widget.TextView;
 
-import com.test.procore.fetchgithubdata.ServiceInterface.IApiService;
+import com.test.procore.fetchgithubdata.R;
+import com.test.procore.fetchgithubdata.serviceinterface.IApiService;
 import com.test.procore.fetchgithubdata.adapter.PRListAdapter;
 import com.test.procore.fetchgithubdata.utils.JsonPojoClass;
 
@@ -20,12 +19,13 @@ import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 
-import static com.test.procore.fetchgithubdata.ServiceInterface.IApiService.GITHUB_BASE_URL;
+import static com.test.procore.fetchgithubdata.serviceinterface.IApiService.GITHUB_BASE_URL;
 
-public class MainActivity extends AppCompatActivity {
+public class GetPRListActivity extends AppCompatActivity {
 
-    private static final String TAG = MainActivity.class.getSimpleName();
+    private static final String TAG = GetPRListActivity.class.getSimpleName();
 
+    private TextView textHeader;
     private RecyclerView recyclerView;
     private RecyclerView.Adapter prListAdapter;
     private RecyclerView.LayoutManager layoutManager;
@@ -35,6 +35,8 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        textHeader   = findViewById(R.id.mainHeaderView);
+
         recyclerView = findViewById(R.id.my_recycler_view);
 
         final Retrofit retrofit = new Retrofit.Builder().baseUrl(GITHUB_BASE_URL)
@@ -48,6 +50,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onResponse(Call<List<JsonPojoClass>> call, Response<List<JsonPojoClass>> response) {
                 List<JsonPojoClass> lists = response.body();
+                textHeader.setText(getString(R.string.title_precursor)+ GITHUB_BASE_URL);
                 setRecyclerView(lists);
             }
             @Override
@@ -60,13 +63,9 @@ public class MainActivity extends AppCompatActivity {
     private void setRecyclerView(List<JsonPojoClass> prList) {
 
         prListAdapter = new PRListAdapter(prList,this);
-        DividerItemDecoration itemDecorator = new DividerItemDecoration(this, DividerItemDecoration.HORIZONTAL);
-
         layoutManager = new LinearLayoutManager(this);
         recyclerView.setLayoutManager(layoutManager);
         recyclerView.setHasFixedSize(true);
-
-        recyclerView.addItemDecoration(itemDecorator);
         recyclerView.setAdapter(prListAdapter);
     }
