+++ app/src/main/res/layout/activity_get_diff.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorAccent"
-    tools:context=".GetDiffActivity">
+    tools:context=".activities.GetDiffActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/diff_recycler_view"
