+++ app/src/main/res/layout/layout_show_differences.xml
@@ -20,12 +20,5 @@
             android:layout_height="wrap_content"
             android:layout_alignParentTop="true"
             android:textStyle="bold" />
-
-        <!--<TextView-->
-            <!--android:id="@+id/textView2"-->
-            <!--android:layout_width="wrap_content"-->
-            <!--android:layout_height="wrap_content"-->
-            <!--android:layout_below="@+id/textView"-->
-            <!--android:layout_marginTop="10dp" />-->
     </RelativeLayout>
 </android.support.v7.widget.CardView>
