+++ app/src/main/res/layout/pull_request_list_item.xml
@@ -2,53 +2,62 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/listRowItem"
     android:layout_width="match_parent"
-    android:layout_height="150dp"
-    android:background="@color/colorlight"
-    android:orientation="vertical">
+    android:layout_height="180dp"
+    android:orientation="vertical"
+    android:padding="10dp">
 
-    <TextView
-        android:id="@+id/layout_title"
-        android:layout_width="match_parent"
-        android:layout_height="32dp"
-        android:gravity="center"
-        android:textSize="25sp"
-        android:textColor="@color/colorPrimary"
-        android:text="" />
 
-    <TextView
-        android:id="@+id/layout_number"
+    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        android:id="@+id/listRowItem1"
         android:layout_width="match_parent"
-        android:layout_height="32dp"
-        android:gravity="center"
-        android:textSize="25sp"
-        android:textColor="@color/colorPrimary"
-        android:text="" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
+        android:layout_height="167dp"
+        android:background="@color/colorlight"
+        android:orientation="vertical"
+        android:padding="10dp">
 
         <TextView
-            android:id="@+id/layout_id"
-            android:layout_width="0dp"
+            android:id="@+id/layout_number"
+            android:layout_width="match_parent"
             android:layout_height="32dp"
-            android:layout_weight="1"
-            android:gravity="center"
-            android:textSize="25sp"
+            android:layout_gravity="start"
+            android:text=""
             android:textColor="@color/colorPrimary"
-            android:text="" />
+            android:textSize="25sp" />
 
         <TextView
-            android:id="@+id/layout_state"
-            android:layout_width="0dp"
-            android:layout_height="32dp"
-            android:textSize="25sp"
+            android:id="@+id/layout_title"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
             android:gravity="center"
-            android:layout_weight="1"
+            android:text=""
             android:textColor="@color/colorPrimary"
-            android:text="" />
+            android:textSize="25sp" />
 
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
+
+            <TextView
+                android:id="@+id/layout_id"
+                android:layout_width="0dp"
+                android:layout_height="32dp"
+                android:layout_weight="1"
+                android:gravity="center"
+                android:text=""
+                android:textColor="@color/colorPrimary"
+                android:textSize="25sp" />
+
+            <TextView
+                android:id="@+id/layout_state"
+                android:layout_width="0dp"
+                android:layout_height="32dp"
+                android:layout_weight="1"
+                android:gravity="center"
+                android:text=""
+                android:textColor="@color/colorPrimary"
+                android:textSize="25sp" />
+
+        </LinearLayout>
     </LinearLayout>
-
 </LinearLayout>
